+++ src/Messages.php
@@ -1,0 +1,25 @@
+<?php
+namespace Alresia\LaravelWassenger;
+
+use Alresia\LaravelWassenger\Traits\WassengerRequest;
+use Alresia\LaravelWassenger\Traits\Messages as MessagesTrait;
+
+/**
+ * @internal
+ */
+class Messages
+{
+
+     /*
+    |--------------------------------------------------------------------------
+    | Messages
+    |--------------------------------------------------------------------------
+    |
+    | Send, browse and manage outbound messages
+    |
+    */
+
+    use MessagesTrait;
+
+    
+}