+++ src/Wassenger.php
@@ -1,8 +1,10 @@
 <?php
 namespace Alresia\LaravelWassenger;
 
-use Alresia\LaravelWassenger\Traits\WassengerRequest;
+use stdClass;
 use Alresia\LaravelWassenger\Traits\Messages;
+use Alresia\LaravelWassenger\Traits\WassengerRequest;
+
 
 /**
  * @internal
@@ -20,6 +22,29 @@ class Wassenger
     */
 
     use WassengerRequest, Messages;
+
+    /**
+     * Search outbound messages, optionally filtered by customer search params. 
+     * 
+     * ************************************************************************
+     * 
+     * @param array|object $request_array
+     * 
+     */
+    public static function numberExist(string $number)
+    {
+
+        if (isset($number)){
+            
+            $data = new stdClass();
+            $data->phone = $number;
+            $instance = new self;
+
+            return $instance->Request(WassengerApiEndpoints::NUMBER_EXIST, $data);
+        }
+           
+
+    }
 
     
 }